cli/pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0" ]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "rdwatch-cli"
version = "0.0.5"
description = "Client for the RD-WATCH server."
license = "Apache-2.0"
authors = [ "Kitware, Inc. <kitware@kitware.com>" ]
readme = "README.md"
repository = "https://github.com/ResonantGeoData/RD-WATCH/django"
keywords = [ "geospatial", "machine learning" ]
classifiers = [
  "Development Status :: 3 - Alpha",
]
packages = [ { include = "rdwatch_cli", from = "src" } ]

[tool.poetry.scripts]
rdwatch = "rdwatch_cli:cli"

[tool.poetry.dependencies]
python = "~3.10.0"
click = "^8.1.3"
rich = "^12.5.1"
Pillow = "^9.2.0"
mercantile = "^1.2.1"
pillow-avif-plugin = "^1.2.2"
aiohttp = "^3.8.3"

[tool.poetry.dev-dependencies]
black = "~22.6.0"
isort = "~5.10.1"
mypy = "0.961"
taskipy = "^1.10.2"
flake8 = "^5.0.4"


[tool.taskipy.tasks]
"lint" = "task lint:black; task lint:isort; task lint:flake8; task lint:mypy"
"lint:fix" = "task lint:black:fix; task lint:isort:fix; task lint:flake8; task lint:mypy;"
"lint:flake8" = "flake8 src"
"lint:black" = "black --check --diff --quiet src"
"lint:black:fix" = "black  --quiet src"
"lint:isort" = "isort --check --diff src"
"lint:isort:fix" = "isort src"
"lint:mypy" = "mypy src"


[tool.isort]
profile = "black"
known_first_party = ["rdwatch_cli"]
sections=["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]


[[tool.mypy.overrides]]
module = "mercantile"
ignore_missing_imports = true
